Move Icons.kt and Theme.kt to latin/utils Move previewDark from Preview.kt to Theme.kt
