+++ app/src/main/java/helium314/keyboard/settings/Preview.kt
@@ -15,5 +15,3 @@ fun initPreview(context: Context) {
     SettingsActivity.settingsContainer = SettingsContainer(context)
     KeyboardIconsSet.instance.loadIcons(context)
 }
-
-const val previewDark = true
