+++ app/src/main/java/helium314/keyboard/settings/SearchScreen.kt
@@ -49,6 +49,9 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import helium314.keyboard.latin.R
+import helium314.keyboard.latin.utils.BackButton
+import helium314.keyboard.latin.utils.CloseIcon
+import helium314.keyboard.latin.utils.SearchIcon
 import helium314.keyboard.settings.preferences.PreferenceCategory
 
 @Composable
