+++ app/src/main/java/helium314/keyboard/settings/WelcomeWizard.kt
@@ -48,7 +48,9 @@ import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
 import helium314.keyboard.latin.R
 import helium314.keyboard.latin.utils.JniUtils
+import helium314.keyboard.latin.utils.Theme
 import helium314.keyboard.latin.utils.UncachedInputMethodManagerUtils
+import helium314.keyboard.latin.utils.previewDark
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
