+++ app/src/main/java/helium314/keyboard/settings/dialogs/ColorThemePickerDialog.kt
@@ -51,14 +51,14 @@ import helium314.keyboard.latin.utils.getStringResourceOrName
 import helium314.keyboard.latin.utils.htmlToAnnotated
 import helium314.keyboard.latin.utils.prefs
 import helium314.keyboard.latin.utils.withHtmlLink
-import helium314.keyboard.settings.DeleteButton
-import helium314.keyboard.settings.EditButton
+import helium314.keyboard.latin.utils.DeleteButton
+import helium314.keyboard.latin.utils.EditButton
 import helium314.keyboard.settings.Setting
 import helium314.keyboard.settings.SettingsActivity
 import helium314.keyboard.settings.SettingsDestination
-import helium314.keyboard.settings.Theme
+import helium314.keyboard.latin.utils.Theme
 import helium314.keyboard.settings.filePicker
-import helium314.keyboard.settings.previewDark
+import helium314.keyboard.latin.utils.previewDark
 import helium314.keyboard.settings.screens.SaveThoseColors
 import helium314.keyboard.settings.contentTextDirectionStyle
 import kotlinx.coroutines.launch
